teste_dot/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
mimesis==18.0.0
sqlparse==0.5.3
pytest==9.0.1
drf-spectacular==0.29.0
coverage==7.12.0
django-nose==1.4.7
